refactor: 페이지별로 선언되던 outlet context 및 navigate 함수화
