+++ src/main/frontend/src/pages/CustomMachines.jsx
@@ -1,18 +1,13 @@
-import React, { useState } from "react";
-import { useOutletContext, useNavigate } from "react-router-dom";
+import React from "react";
 import CustomProducts from "./../components/CustomProducts";
 import styles from "./CustomMachines.module.css";
 import { useMachines } from "./../hooks/useProducts";
 import Banner from "./../components/Banner";
 import Button from "../components/Button";
+import { useGoHome } from "../hooks/useNavigator";
 
 export default function InputMachines() {
-  const { handleHidden } = useOutletContext();
-  const navigate = useNavigate();
-  const handleClick = () => {
-    handleHidden();
-    navigate("/");
-  };
+  const { handleClick } = useGoHome();
 
   const {
     productsQuery: { isLoading, error, data: machines },
